moved test to be AndroidTest (complaining about bocking)
